Introduce ContractStateService interface
