+++ docs/configuration.md
@@ -691,14 +691,14 @@ The following table lists the available properties along with their default valu
 value, it is recommended to only populate overridden properties in the custom `application.yml`.
 
 | Name                                                         | Default                                            | Description                                                                                                                                                                                      |
-|--------------------------------------------------------------|----------------------------------------------------| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
+|--------------------------------------------------------------|----------------------------------------------------|--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
 | `hiero.mirror.web3.cache.contract`                           | expireAfterAccess=60m,maximumSize=1000,recordStats | Cache configuration for contract                                                                                                                                                                 |
 | `hiero.mirror.web3.cache.contractState`                      | expireAfterWrite=2s,maximumSize=25000,recordStats  | Cache configuration for contract state                                                                                                                                                           |
 | `hiero.mirror.web3.cache.contractSlots`                      | expireAfterWrite=5m,maximumSize=3000,recordStats   | Cache for contract slots queried within a findStorage query                                                                                                                                      |
 | `hiero.mirror.web3.cache.enableBatchContractSlotCaching `    | true                                               | Flag enabling batch contract slot caching                                                                                                                                                        |
 | `hiero.mirror.web3.cache.entity `                            | expireAfterWrite=1s,maximumSize=10000,recordStats  | Cache configuration for entity                                                                                                                                                                   |
 | `hiero.mirror.web3.cache.fee`                                | expireAfterWrite=10m,maximumSize=20,recordStats    | Cache configuration for fee related info                                                                                                                                                         |
-| `hiero.mirror.web3.cache.maxSlotsPerContract`                | 1500                                               | The amount of slot keys that are cached per contract. Used for batch calls.                                                                                                                      |
+| `hiero.mirror.web3.cache.slotsPerContract`                   | expireAfterWrite=5m,maximumSize=1500               | Cache for slot keys per contract.Separate instance used for the value of each entry of the contractSlots cache                                                                                   |
 | `hiero.mirror.web3.cache.token`                              | expireAfterWrite=1s,maximumSize=10000,recordStats  | Cache configuration for token related info                                                                                                                                                       |
 | `hiero.mirror.web3.db.host`                                  | 127.0.0.1                                          | The IP or hostname used to connect to the database                                                                                                                                               |
 | `hiero.mirror.web3.db.name`                                  | mirror_node                                        | The name of the database                                                                                                                                                                         |
