+++ web3/src/main/java/org/hiero/mirror/web3/state/Utils.java
@@ -3,7 +3,6 @@
 package org.hiero.mirror.web3.state;
 
 import static org.hiero.mirror.common.util.DomainUtils.isLongZeroAddress;
-import static org.hiero.mirror.common.util.DomainUtils.leftPadBytes;
 
 import com.hedera.hapi.node.base.Key;
 import com.hedera.hapi.node.base.KeyList;
@@ -11,9 +10,7 @@ import com.hedera.hapi.node.base.Timestamp;
 import com.hedera.pbj.runtime.ParseException;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import java.time.Instant;
-import java.util.Objects;
 import lombok.experimental.UtilityClass;
-import org.apache.tuweni.bytes.Bytes32;
 import org.hyperledger.besu.datatypes.Address;
 
 @UtilityClass
@@ -49,16 +46,5 @@ public class Utils {
 
     public static boolean isMirror(final Address address) {
         return address != null && isLongZeroAddress(address.toArrayUnsafe());
-    }
-
-    /**
-     * Converts a byte array to a 32-length left-padded PBJ Bytes object.
-     *
-     * @param byteArray The byte array.
-     * @return The PBJ Bytes object.
-     */
-    public static Bytes convertToLeftPaddedBytes(final byte[] byteArray) {
-        return Bytes.wrap(
-                leftPadBytes(Objects.requireNonNullElseGet(byteArray, Bytes.EMPTY::toByteArray), Bytes32.SIZE));
     }
 }
